grpc: move symbols proto definitions to internal/symbols/proto
