+++ internal/symbols/proto/BUILD.bazel
