+++ internal/symbols/proto/buf.gen.yaml
