+++ internal/symbols/proto/conversion.go
