+++ internal/symbols/proto/conversion_test.go
