+++ internal/symbols/proto/doc.go
@@ -1,0 +1,2 @@
+// Package proto contains protocol buffer definitions for the symbols service.
+package proto
